mid-work commit for #55

+++ src/main/java/ch/uzh/ifi/hase/soprafs24/model/game/Turn.java
@@ -7,10 +7,10 @@ import java.util.List;
 
 @Getter
 public class Turn {
-    private final int userId;
+    private final long userId;
     private final List<Integer> picks = new ArrayList<>();
 
-    public Turn(int userId) {
+    public Turn(long userId) {
         this.userId = userId;
     }
